test/crash_fuzzing_agent/requirements.txt
openai_agents
python-dotenv
requests
sqlglot
snowflake-connector-python==3.15.0